move content_helper
